+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - new: `Timestamp` wrapper around timestamps
  - update: `Timespan`: get durations with specified units only
  - update: `Timespan`: add month unit
+ - update: `Timespan`: **symbolType** renamed to **symbolFormat**
 
 ### 0.1.0 (2022 Aug 13)